pyproject.toml
[tool.poetry]
name = "libraryapi"
version = "0.1.0"
description = ""
authors = ["ahmedkhairym0 <ahmedkhairym0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.5"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
